refactor: converted main js files to ts
