+++ src/shims-vue.d.ts
@@ -1,0 +1,5 @@
+declare module '*.vue' {
+  import { defineComponent } from 'vue'
+  const Component: ReturnType<typeof defineComponent>
+  export default Component
+}
